Upgrade to latest axle drivers and fix resulting compilation error
This means we're now using Long as the type for generated ids, but that's
what we've always recommended anyway.

+++ hibernate-rx-core/src/main/java/org/hibernate/rx/impl/PoolConnection.java
@@ -58,7 +58,7 @@ public class PoolConnection implements RxConnection {
 	}
 
 	@Override
-	public CompletionStage<Optional<Integer>> updateReturning(String sql, Tuple parameters) {
+	public CompletionStage<Optional<Long>> updateReturning(String sql, Tuple parameters) {
 		return preparedQuery( sql, parameters )
 				.thenApply( rows -> Optional.ofNullable( rows.property(MySQLClient.LAST_INSERTED_ID) ) );
 	}
